+++ CHANGELOG.md
@@ -35,6 +35,12 @@ Section Order:
 ### Security
 -->
 
+### Changed
+
+- `views.ajax_dashboard` refactored
+- Dashboard buttons to Font Awesome icons
+- Dashboard JS refactored
+
 ## [2.6.1] - 2025-08-19
 
 ### Fixed